--- index.ts
@@ -1,5 +1,0 @@
-import { MainApp } from "./src/mainApp";
-
-(async() => {
-    const application = await MainApp.createClassInstance();
-})();
